Temporarily removed diagonal movement
Now there is no problem with collision. If we get time over we can
implement diagonal movement.

+++ Player.h
@@ -17,7 +17,7 @@ public:
 	float GetY();
 	bool IsVisible();
 	EEntityType GetType();
-	void SetPosition(float x, float y);
+	void SetPosition(int x, int y);
 	float GetDirectionX();
 	float GetDirectionY();
